+++ src/Jobs/QueueExport.php
@@ -6,6 +6,7 @@ use Illuminate\Contracts\Queue\ShouldQueue;
 use Illuminate\Foundation\Bus\Dispatchable;
 use Maatwebsite\Excel\Concerns\WithMultipleSheets;
 use Maatwebsite\Excel\Files\TemporaryFile;
+use Maatwebsite\Excel\Jobs\Middleware\LocalizeJob;
 use Maatwebsite\Excel\Writer;
 use Throwable;
 
@@ -57,21 +58,23 @@ class QueueExport implements ShouldQueue
      */
     public function handle(Writer $writer)
     {
-        $writer->open($this->export);
+        (new LocalizeJob($this->export))->handle($this, function () use ($writer) {
+            $writer->open($this->export);
 
-        $sheetExports = [$this->export];
-        if ($this->export instanceof WithMultipleSheets) {
-            $sheetExports = $this->export->sheets();
-        }
+            $sheetExports = [$this->export];
+            if ($this->export instanceof WithMultipleSheets) {
+                $sheetExports = $this->export->sheets();
+            }
 
-        // Pre-create the worksheets
-        foreach ($sheetExports as $sheetIndex => $sheetExport) {
-            $sheet = $writer->addNewSheet($sheetIndex);
-            $sheet->open($sheetExport);
-        }
+            // Pre-create the worksheets
+            foreach ($sheetExports as $sheetIndex => $sheetExport) {
+                $sheet = $writer->addNewSheet($sheetIndex);
+                $sheet->open($sheetExport);
+            }
 
-        // Write to temp file with empty sheets.
-        $writer->write($sheetExport, $this->temporaryFile, $this->writerType);
+            // Write to temp file with empty sheets.
+            $writer->write($sheetExport, $this->temporaryFile, $this->writerType);
+        });
     }
 
     /**