+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Added
+
+- Added an ability to prepare rows before appending rows to sheet. Just add `prepareRows` method for your export class if needed.
+
 ## [3.1.24] - 2020-10-28
 
 ### Added