+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 
 - Added an ability to prepare rows before appending rows to sheet. Just add `prepareRows` method for your export class if needed.
 - Added ab ability to catch exceptions from `QueueExport` job. Just add `failed` method for your export class if needed.
+- Added ab ability to set locale for queued export. Just implement `Illuminate\Contracts\Translation\HasLocalePreference` for your export.
 
 ## [3.1.24] - 2020-10-28